JobAssistant-main/Backend/requirements.txt
pypdf==5.4.0
python-docx==1.1.2
fastapi==0.116.1
httpx==0.28.1
pydantic-settings==2.10.1
supabase==2.18.1
uvicorn[standard]==0.35.0